fastApi/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.0
h11==0.14.0
idna==3.4
loguru==0.6.0
psycopg2-binary==2.9.5
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
uvicorn==0.21.1
nltk
deep_translator
pyspellchecker